Tidy up all files

+++ lib/labelvalidator/vcs.rb
@@ -5,25 +5,31 @@ require 'octokit'
 module LabelValidator
   # Used to handle calls to VCS
   class Vcs
-    def initialize(token:, pull_request:)
+    def initialize(token:, pull_request:, comments_enabled: ENV['COMMENTS_ENABLED'])
       @client = Octokit::Client.new(access_token: token)
       @pull_request = pull_request
       @comment_base = 'LabelValidator: This will be bumped on merge by a'
+      @comments_enabled = comments_enabled
+    end
+
+    def default_branch_target?
+      true if @pull_request['base']['ref'] == @pull_request['base']['repo']['default_branch']
+      false
     end
 
     def status_check(state:)
       raise ArgumentError, 'State must be pending, success, failure' unless %w(pending success failure).include?(state)
 
       @client.create_status(@pull_request['head']['repo']['full_name'],
-        @pull_request['head']['sha'],
-        state,
-        { :context => 'Release Label Validator',
-          :description => 'Checking this commit has a single "Release: Major|Minor|Patch" label'
-        }
-      )
+                            @pull_request['head']['sha'],
+                            state,
+                            { context: 'Release Label Validator',
+                              description: 'Checking this commit has a single "Release: Major|Minor|Patch" label' })
     end
 
-    def delete_all_bot_comments()
+    def delete_all_bot_comments
+      return 'skip comments' unless @comments_enabled
+
       bot_comments = get_bot_comments(@comment_base)
       bot_comments.each do |comment|
         delete_comment(comment)
@@ -31,12 +37,14 @@ module LabelValidator
     end
 
     def bot_comment(release_level)
+      return 'skip comments' unless @comments_enabled
+
       @comment_base = 'LabelValidator: This will be bumped on merge by a'
       comment_full = "#{@comment_base} #{release_level} version"
 
       bot_comments = get_bot_comments(@comment_base)
-      if bot_comments.count > 0
-          update_bot_comments(bot_comments, comment_full)
+      if bot_comments.count.positive?
+        update_bot_comments(bot_comments, comment_full)
       else
         add_bot_comment(comment_full)
       end
@@ -46,17 +54,16 @@ module LabelValidator
 
     def get_bot_comments(filter)
       user = @client.user[:login]
-      all_comments =  @client.issue_comments(@pull_request['head']['repo']['full_name'], @pull_request['number'])
+      all_comments = @client.issue_comments(@pull_request['head']['repo']['full_name'], @pull_request['number'])
       bot_comments = all_comments.select { |c| c[:user][:login] == user }
       filter_regex = Regexp.new(filter)
-      return bot_comments.select { |c| c[:body] =~ filter_regex }
+      bot_comments.select { |c| c[:body] =~ filter_regex }
     end
 
     def add_bot_comment(body)
       @client.add_comment(@pull_request['head']['repo']['full_name'],
-        @pull_request['number'],
-        body
-      )
+                          @pull_request['number'],
+                          body)
     end
 
     def update_bot_comments(comments, body)
@@ -76,8 +83,7 @@ module LabelValidator
 
     def delete_comment(comment)
       @client.delete_comment(@pull_request['head']['repo']['full_name'],
-        comment['id']
-      )
+                             comment['id'])
     end
   end
 end
